Final wrap-up of refactoring OAuthClient
Closes #37225
