+++ tests/utils-shared/src/main/java/org/keycloak/testsuite/util/oauth/PasswordGrantRequest.java
@@ -2,7 +2,6 @@ package org.keycloak.testsuite.util.oauth;
 
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.keycloak.OAuth2Constants;
-import org.keycloak.constants.AdapterConstants;
 import org.keycloak.util.TokenUtil;
 
 import java.io.IOException;
